algorithm/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
ale-py==0.7.5
alembic==1.9.1
ansicon==1.89.0
anytree==2.8.0
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
backcall==0.2.0
bleach==6.0.0
blessed==1.19.1
build==0.10.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cliff==4.1.0
cloudpickle==2.0.0
cmaes==0.9.0
cmd2==2.4.2
colorama==0.4.4
colorful==0.5.5
colorlog==6.7.0
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.0
distlib==0.3.6
dm-tree==0.1.7
docutils==0.19
exceptiongroup==1.0.0
filelock==3.8.2
fonttools==4.28.5
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.57.0
gpustat==1.0.0
greenlet==2.0.1
grpcio==1.50.0
gym==0.21.0
-e git+https://github.com/BTHeinbach/gym-flp.git@371aad2e14a833226a5f79721e8f3efe143a7d0c#egg=gym_flp
gym-notices==0.0.8
idna==3.4
imageio==2.13.5
importlib-metadata==4.13.0
importlib-resources==5.10.1
iniconfig==1.1.1
ipython==7.31.0
jaraco.classes==3.2.3
jedi==0.18.1
jinxed==1.2.0
jsonschema==4.17.3
keyring==23.13.1
kiwisolver==1.3.2
lz4==4.0.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.3
networkx==2.8.8
numpy==1.23.4
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencensus==0.11.0
opencensus-context==0.1.3
optuna==3.0.5
packaging==21.3
pandas==1.5.1
parso==0.8.3
pbr==5.11.0
pickleshare==0.7.5
Pillow==9.0.0
pkginfo==1.9.6
platformdirs==2.6.0
plotly==5.11.0
pluggy==1.0.0
prettytable==3.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.24
protobuf==3.19.6
psutil==5.9.4
py-spy==0.3.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pygame==2.1.2
Pygments==2.11.2
pyocclient==0.6
pyparsing==3.0.6
pyperclip==1.8.2
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.3
PyWavelets==1.4.1
pywin32-ctypes==0.2.0
PyYAML==6.0
ray @ https://s3-us-west-2.amazonaws.com/ray-wheels/latest/ray-3.0.0.dev0-cp310-cp310-win_amd64.whl
readme-renderer==37.3
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
rsa==4.9
scikit-image==0.19.3
scipy==1.8.1
shapely==2.0.0
six==1.16.0
smart-open==6.2.0
SQLAlchemy==1.4.45
stable-baselines3==1.6.2
stevedore==4.1.1
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tifffile==2022.10.10
tomli==2.0.1
torch==1.12.1
tqdm==4.64.1
traitlets==5.1.1
twine==4.0.2
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.17.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.9.0
